create chapter 02 booleans
