+++ exercises/src/01.Arithmetic/arith0.js
@@ -1,14 +1,10 @@
-import { nav1Html } from './nav1'
-import { menuHtml, setMenu } from '../menu'
-import { arith1Html } from './arith1'
-import { arith2Html } from './arith2'
-import { arith3Html } from './arith3'
+import { setNav1 } from './nav1'
 
 export function arith0(content, setNav) {
     content.innerHTML = arith0Html()
     
     if(setNav) {
-        setNavigation(content)
+        setNav1(content)
     }
 }
 
@@ -34,31 +30,4 @@ console.log("El resultado de la suma es: " + resulado.toFixed(2))
     </pre>
 </details>
 `
-}
-
-async function setNavigation(content) {
-
-    document.querySelector("#nav").innerHTML = nav1Html()
-    document.querySelector("#menu").innerHTML = menuHtml()
-    setMenu()
-    
-    document.querySelector("#arith0").addEventListener('click', () =>
-    {
-        arith0(content, false)
-    })
-
-    document.querySelector("#arith1").addEventListener('click', () =>
-    {
-        content.innerHTML = arith1Html()
-    })
-    
-    document.querySelector("#arith2").addEventListener('click', () =>
-    {
-        content.innerHTML = arith2Html()
-    })
-
-    document.querySelector("#arith3").addEventListener('click', () =>
-    {
-        content.innerHTML = arith3Html()
-    })
 }
