+++ exercises/src/01.Arithmetic/nav1.js
@@ -1,3 +1,9 @@
+import { menuHtml, setMenu } from '../menu'
+import { arith0 } from './arith0'
+import { arith1Html } from './arith1'
+import { arith2Html } from './arith2'
+import { arith3Html } from './arith3'
+
 export function nav1Html() {
     return `
 <ul class="nav">
@@ -8,4 +14,31 @@ export function nav1Html() {
     <li><a class="tab" href="#" id="arith3">Ejercicio 3</a></li>
 </ul>
 `
+}
+
+export function setNav1(content) {
+
+    document.querySelector("#nav").innerHTML = nav1Html()
+    document.querySelector("#menu").innerHTML = menuHtml()
+    setMenu()
+    
+    document.querySelector("#arith0").addEventListener('click', () =>
+    {
+        arith0(content, false)
+    })
+
+    document.querySelector("#arith1").addEventListener('click', () =>
+    {
+        content.innerHTML = arith1Html()
+    })
+    
+    document.querySelector("#arith2").addEventListener('click', () =>
+    {
+        content.innerHTML = arith2Html()
+    })
+
+    document.querySelector("#arith3").addEventListener('click', () =>
+    {
+        content.innerHTML = arith3Html()
+    })
 }
