--- exercises/src/02.Strings/string0.js
@@ -1,9 +1,0 @@
-export function string0Html() {
-    //TODO: Excercise 1 -> Get String Length
-    //TODO: Excercise 2 -> Concatenate strings & format with ${}
-    //TODO: Excercise 3 -> Upper & Lower
-    //TODO: Excercise 4 -> Trim
-    //TODO: Excercise 5 -> Replace & Slice
-    //TODO: Excercise 6 -> Revert & Palindrome
-    return `<h2>Esto es el Capitulo 02 - Manejo de Strings</h2><h3>Ejercicio 0</h3><p>Neque porro quisquam est qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit...</p>`
-}
