+++ exercises/src/03.Strings/string0.js
@@ -1,0 +1,9 @@
+export function string0Html() {
+    //TODO: Excercise 1 -> Get fixed length
+    //TODO: Excercise 2 -> Get variable length with prompt
+    //TODO: Excercise 2 -> Concatenate strings & format with ${}
+    //TODO: Excercise 3 -> Upper, Lower & Trim
+    //TODO: Excercise 4 -> Replace & Slice
+    //TODO: Excercise 5 -> Revert & Palindrome
+    return `<h2>Capitulo 03 - Manejo de Strings</h2><h3>Ejercicio 0</h3><p>Neque porro quisquam est qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit...</p>`
+}
